+++ package.json
@@ -13,6 +13,7 @@
     "build:bundle": "lerna run build:bundle --stream",
     "build:app": "lerna run pack --stream && cd test/apps/vanilla && rm -rf node_modules && yarn install --no-immutable && yarn build",
     "build:react-app": "lerna run pack --stream && cd test/apps/react && rm -rf node_modules && yarn install --no-immutable && yarn build",
+    "build:test-extension": "lerna run pack --stream && cd test/apps/extension && rm -rf node_modules && yarn install --no-immutable && yarn build",
     "build:apps": "yarn build:app && yarn build:react-app",
     "format": "prettier --check .",
     "lint": "scripts/cli lint .",
