+++ test/apps/extension/package.json
@@ -8,6 +8,12 @@
   "dependencies": {
     "@datadog/browser-rum": "file:../../../packages/rum/package.tgz"
   },
+  "resolutions": {
+    "@datadog/browser-rum-core": "file:../../../packages/rum-core/package.tgz",
+    "@datadog/browser-core": "file:../../../packages/core/package.tgz",
+    "@datadog/browser-rum-slim": "file:../../../packages/rum-slim/package.tgz",
+    "@datadog/browser-worker": "file:../../../packages/worker/package.tgz"
+  },
   "devDependencies": {
     "webpack": "5.94.0",
     "webpack-cli": "6.0.1"
