+++ test/e2e/scenario/developer-extension/developerExtension.scenario.ts
@@ -1,35 +1,14 @@
 import path from 'path'
-import { test as base, chromium, expect } from '@playwright/test'
-import type { Page, BrowserContext } from '@playwright/test'
+import { expect } from '@playwright/test'
+import type { Page } from '@playwright/test'
+import { createExtensionTest } from '../utils/extensionFixture'
 
-const test = base.extend<{
-  context: BrowserContext
-  extensionId: string
+const pathToExtension = path.join(process.cwd(), 'developer-extension', 'dist')
+const test = createExtensionTest(pathToExtension).extend<{
   developerExtension: DeveloperExtensionPage
 }>({
-  // eslint-disable-next-line no-empty-pattern
-  context: async ({}, use) => {
-    const pathToExtension = path.join(process.cwd(), 'developer-extension', 'dist')
-
-    const context = await chromium.launchPersistentContext('', {
-      channel: 'chromium',
-      args: [`--disable-extensions-except=${pathToExtension}`, `--load-extension=${pathToExtension}`],
-    })
-    await use(context)
-    await context.close()
-  },
-  extensionId: async ({ context }, use) => {
-    let [background] = context.serviceWorkers()
-    if (!background) {
-      background = await context.waitForEvent('serviceworker')
-    }
-
-    const extensionId = background.url().split('/')[2]
-    await use(extensionId)
-  },
   developerExtension: async ({ page, extensionId }, use) => {
     await page.goto(`chrome-extension://${extensionId}/panel.html`)
-
     await use(new DeveloperExtensionPage(page))
   },
 })
